restructure plugins
