+++ .babelrc
@@ -1,3 +1,6 @@
 {
-  "presets": ["env", "stage-2"]
+  "presets": [
+    ["env", { "useBuiltIns": true }],
+    "stage-2"
+  ]
 }
