+++ src/plugins/plugin-leaflet/carto-layer-leaflet.js
@@ -1,5 +1,7 @@
 import Promise from 'bluebird';
-import cartoService from './carto-layer-service';
+import cartoService from '../../services/carto-layer-service';
+
+const { L } = window;
 
 const CartoLayer = (layerModel) => {
   const { layerConfig, interactivity } = layerModel;
