+++ src/services/carto-layer-service.js
@@ -1,5 +1,5 @@
 import Promise from 'bluebird';
-import { get } from '../../helpers';
+import { get } from '../helpers';
 
 const cartoLayerService = (layerModel) => {
   const { layerConfig, interactivity } = layerModel;
